movwe changes to node file
